Core data: add a private media "edit" action to ensure cache invalidation (#70814)
This commit adds a new private action `editMediaEntity` to Core Data and associated unit tests. Because it needs to be called in block-editor hooks, the action is accessible in block editor settings via a private key. The primary motivation is to provide a core data integration for the `WP_REST_Attachments_Controller` REST API's `wp/v2/media/{ id }/edit` route. 


Co-authored-by: ramonjd <ramonopoly@git.wordpress.org>
Co-authored-by: youknowriad <youknowriad@git.wordpress.org>
Co-authored-by: talldan <talldanwp@git.wordpress.org>
Co-authored-by: andrewserong <andrewserong@git.wordpress.org>