Backport: Avoid warning for non-string class attributes.
Trac ticket: Core-59622

This is a backport of WordPress/wordpress-develop#5486, which
introduces an additional check for non-string values of the
`class` attribute when rendering background support for blocks.

This check avoids issuing a warning when a boolean `true` value
of the attribute is passed on to `str_ends_with()`.

Co-authored-by: Tonya Mork <hellofromTonya@git.wordpress.org>
Co-authored-by: Peter Wilson <peterwilsoncc@git.wordpress.org>
Co-authored-by: Jon Surrell <jonsurrell@git.wordpress.org>
Co-authored-by: George Mamadashvili <mamaduka@git.wordpress.org>
Github-PR: 71594
Github-PR-URL: https://github.com/WordPress/gutenberg/pull/71594
Branch-name: backport/block-supports-background-non-string-attribute-check
Trac-ticket: 59622