Allow submenu flyout menus to be visible outside of template part canvas (#72013)
When editing a navigation template part or thin template part with a navigation (like a header), the submenus dropdown menus are contained within the small area of the iframe and blocked from overflowing. This makes the submenus unusable as they are cut off. This commit expands the iframe to be the full size of the canvas, and fakes a background on the iframe HTML element by hardcoding the #ddd color as a background. Then, the body has the color of the user's chosen background and is centered using inlined CSS.

---------

Co-authored-by: Ben Dwyer <ben@scruffian.com>