Workflows: Performance: Fix RegExp for wpVersion option (#72327)
Follow-up of #32244 (4½ years later!)

The original regular expression was missing a backslash to match a
literal period (\.), and thus it could match any character.

In practice, this should never matter because of WordPress's versioning
conventions. However, if we ever reach a x.100 minor release, the CI
performance tests would fail. ;)

Co-authored-by: mcsf <mcsf@git.wordpress.org>
Co-authored-by: youknowriad <youknowriad@git.wordpress.org>